# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contacts
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Contact list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactBody"
      responses:
        "201":
          description: Contact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
...
  /deals:
    get:
      operationId: listDeals
      tags: [deals]
      summary: List all deals
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: stage
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Deal list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
    post:
      operationId: createDeal
      tags: [deals]
      summary: Create a deal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealBody"
      responses:
        "201":
          description: Deal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
...
  /artists:
    get:
      operationId: listArtists
      tags: [artists]
      summary: List all artists
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: leadTier
          in: query
          schema:
            type: string
        - name: outreachStatus
          in: query
          schema:
            type: string
        - name: genre
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: followersEstimate
          in: query
          schema:
            type: string
        - name: engagementLevel
          in: query
          schema:
            type: string
        - name: labelStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of artists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artist"
    post:
      operationId: createArtist
      tags: [artists]
      summary: Create artist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtistBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
...
  /royalties:
    get:
      operationId: listRoyalties
      tags: [royalties]
      summary: List royalty records
      security:
        - bearerAuth: []
      parameters:
        - name: artistId
          in: query
          schema:
            type: integer
        - name: releaseId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of royalties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Royalty"
    post:
      operationId: createRoyalty
      tags: [royalties]
      summary: Create a royalty record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoyaltyBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Royalty"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ContactBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        organization:
          type: string
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
      required: [name]
...
    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        company:
          type: string
          nullable: true
        organization:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, tags, createdAt, updatedAt]
...
    DealBody:
      type: object
      properties:
        title:
          type: string
        value:
          type: number
          nullable: true
        stage:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
          nullable: true
        expectedCloseDate:
          type: string
          format: date
          nullable: true
        contactId:
          type: integer
          nullable: true
        assignedTo:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required: [title]
    Deal:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        value:
          type: number
          nullable: true
        stage:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        expectedCloseDate:
          type: string
          format: date
          nullable: true
        contactId:
          type: integer
          nullable: true
        contactName:
          type: string
          nullable: true
        assignedTo:
          type: integer
          nullable: true
        assignedToName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        closedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, stage, priority, createdAt, updatedAt]
...
    ArtistBody:
      type: object
      properties:
        name:
          type: string
        genre:
          type: string
          nullable: true
        labelStatus:
          type: array
          items:
            type: string
        bio:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        outreachStatus:
          type: string
          enum: [new, contacted, in_talks, signed, passed]
        revenuePotential:
          type: string
          nullable: true
        followersEstimate:
          type: string
          nullable: true
        engagementLevel:
          type: string
          enum: [low, medium, high]
          nullable: true
        streamingLinks:
          type: object
          additionalProperties:
            type: string
        socialLinks:
          type: object
          additionalProperties:
            type: string
        tags:
          type: array
          items:
            type: string
        contactId:
          type: integer
          nullable: true
        spotifyId:
          type: string
          nullable: true
        youtubeChannelId:
          type: string
          nullable: true
        originCity:
          type: string
          nullable: true
        originState:
          type: string
          nullable: true
        originCountry:
          type: string
          nullable: true
        photoUrls:
          type: array
          items:
            type: string
      required: [name]
    Artist:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        genre:
          type: string
          nullable: true
        labelStatus:
          type: array
          items:
            type: string
        bio:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        lat:
          type: number
          format: float
          nullable: true
        lng:
          type: number
          format: float
          nullable: true
        outreachStatus:
          type: string
        revenuePotential:
          type: string
          nullable: true
        followersEstimate:
          type: string
          nullable: true
        engagementLevel:
          type: string
          enum: [low, medium, high]
          nullable: true
        imageUrl:
          type: string
          nullable: true
        spotifyId:
          type: string
          nullable: true
        youtubeChannelId:
          type: string
          nullable: true
        originCity:
          type: string
          nullable: true
        originState:
          type: string
          nullable: true
        originCountry:
          type: string
          nullable: true
        photoUrls:
          type: array
          items:
            type: string
        streamingLinks:
          type: object
          additionalProperties:
            type: string
        socialLinks:
          type: object
          additionalProperties:
            type: string
        tags:
          type: array
          items:
            type: string
        contactId:
          type: integer
          nullable: true
        aiAnalysis:
          $ref: "#/components/schemas/ArtistAiAnalysis"
          nullable: true
        createdBy:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, labelStatus, outreachStatus, streamingLinks, socialLinks, tags, createdBy, createdAt, updatedAt]
    ArtistAiAnalysis:
      type: object
      properties:
        id:
          type: integer
        artistId:
          type: integer
        summary:
          type: string
        brandingScore:
          type: integer
        growthScore:
          type: integer
        professionalismScore:
          type: integer
        leadTier:
          type: string
        recommendations:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, artistId, summary, brandingScore, growthScore, professionalismScore, leadTier, recommendations, createdAt, updatedAt]
...
    Royalty:
      type: object
      properties:
        id:
          type: integer
        artistId:
          type: integer
        releaseId:
          type: integer
          nullable: true
        periodStart:
          type: string
        periodEnd:
          type: string
        streamCount:
          type: integer
        grossCents:
          type: integer
        netCents:
          type: integer
        splitPct:
          type: integer
        status:
          type: string
          enum: [pending, processing, paid, disputed]
        paidAt:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, artistId, periodStart, periodEnd, streamCount, grossCents, netCents, splitPct, status, createdAt]
    RoyaltyBody:
      type: object
      properties:
        artistId:
          type: integer
        releaseId:
          type: integer
        periodStart:
          type: string
        periodEnd:
          type: string
        streamCount:
          type: integer
        grossCents:
          type: integer
        netCents:
          type: integer
        splitPct:
          type: integer
        status:
          type: string
          enum: [pending, processing, paid, disputed]
        notes:
          type: string
      required: [artistId, periodStart, periodEnd, streamCount, grossCents, netCents, splitPct]
...